+++ Tests/Runtime/Unity.Properties.Tests/TestData/FormerlySerializedTypes.cs
@@ -1,5 +1,4 @@
 using System.Collections.Generic;
-using UnityEngine.Serialization;
 
 namespace Unity.Properties.Tests
 {
@@ -7,14 +6,17 @@ namespace Unity.Properties.Tests
     {
         public float x;
         public float y;
+        [Property] public float z { get; set; }
     }
     
     public struct MyOwnVectorWithFormerlySerializedAs
     {
-        [FormerlySerializedAs("x")]
-        public float X;
+        [UnityEngine.Serialization.FormerlySerializedAs("x")]
+        public float RenamedX;
         [FormerlySerializedAs("y")]
-        public float Y;
+        public float RenamedY;
+        [FormerlySerializedAs("z")]
+        [Property] public float RenamedZ { get; set; }
     }
     
     public class FormerlySerializedAsMockData
@@ -22,12 +24,14 @@ namespace Unity.Properties.Tests
         public float MyFloat;
         public List<int> SomeList;
         public MyOwnVector MyVector;
+        public string SomeString { get; set; }
     }
     
     public class FormerlySerializedAsData
     {
-        [FormerlySerializedAs("MyFloat")]
+        [UnityEngine.Serialization.FormerlySerializedAs("MyFloat")]
         public float SomeSimpleFloat = 0;
+
         [FormerlySerializedAs("TryToFoolYou")]
         [FormerlySerializedAs("SomeList")]
         public List<int> ListOfInts = new List<int>( );