+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 All notable changes to this package will be documented in this file.
 
+## [0.10.4-preview] - 2020-01-13
+### Added
+* Added new `FormerlySerializedAs` attribute that can be applied to struct, class, field and property types.
+* Added new methods `TypeConstruction.[Try]ConstructFromAssemblyQualifiedTypeName` that can construct types from assembly qualified type name. If the assembly qualified type name fails to resolve to a valid type, it will try from types marked with `FormerlySerializedAs` that matches.
+
 ## [0.10.3-preview] - 2019-11-08
 ### Fixed
 * AOT Fix: Allows for registering container types with the internal generic virtual calls in Properties